Move negative bosons spline into the .cpp files (#182)
* Fix

* Fix windows
